refactor(schema): Remove Aurelia-Dependency-Injection
Progress update: Testing indicates that Aurelia Dependency Injection is not working out the way I would like it to, and overall seems too heavily tied to the rest of the Aurelia framework to be a drop-in solution. It does however seem that this refactor may be possible if I can find the right IoC library to keep from generating duplicate type definitions. Next step will be to try and adopt an alternate solution.

Added Babel-Plugin-Import-Glob to replace require loops in a future update.