refactor(schema): Add Printing Type/Model, Condense Queries/Mutations
Fairly significant refactor of all types. Major changes include:

- Abstracting out Create/Read/Update/Destroy methods for most Queries/Mutations across all types. Some exceptions remain for types which need to create additional entries upon creation.

- Added logging support to most operations, improving the ability to debug errors while running populate.

- Splute up Card, creating the Printing type. A Card should ever only be created once, whereas a Printing should hold all information specific to that individual printing of a card in a particular set, etc.

- Refactored parts of populate to run non-dependent inserts such as Card Type data in batch mutations. This reduces the number of network round trips required, allowing data to be added/updated in bulk.

A major focus of this cycle was to improve performance and debugging of the application. As such some additional tools have been added, such as Webpack to optionally build and bundle the server. For the time being this is not the recommended way to run the server. Performance improvements are a work in progress and the next steps will involve integrating Dataloader and optimizing SQL queries.